set ideal module
