+++ ot/mpz-ot-core/src/ideal/mod.rs
@@ -1,0 +1,3 @@
+//! Define ideal functionalities of OT.
+
+pub mod ideal_cot;
